+++ packages/react/src/components/Box/Box.test.tsx
@@ -18,9 +18,6 @@ describe('Box', () => {
     const box = screen.getByTitle('box');
 
     expect(box.classList).toContain('box');
-    expect(box.classList).toContain('mediumShadow');
-    expect(box.classList).toContain('mediumBorderRadius');
-    expect(box.classList).toContain('defaultBackground');
   });
 
   it('should render a div with correct classname when shadow is xsmall', () => {
@@ -42,12 +39,5 @@ describe('Box', () => {
     const box = screen.getByTitle('box');
 
     expect(box.classList).toContain('smallBorderRadius');
-  });
-
-  it('should render the box as a header', () => {
-    render({ as: 'header' });
-    const box = screen.getByTitle('box');
-
-    expect(box.tagName).toBe('HEADER');
   });
 });