+++ packages/react/src/components/Combobox/test-data/getCountryName.ts
@@ -1,0 +1,4 @@
+import { countries } from './countries';
+
+export const getCountryName = (code: string): string =>
+  code in countries ? countries[code as keyof typeof countries] : '';